backend/requirements.txt
flask==3.0.2
flask-cors==4.0.0
kafka-python-ng==2.2.3  # Using the newer kafka-python-ng instead of kafka-python
psycopg2-binary==2.9.9
python-dotenv==1.0.1
requests==2.31.0
six==1.16.0
ollama==0.1.6